refactor: add BoundPredicateEvaluator, use for ManifestEvaluator, move inclusive projection and manifest evaluator into own files
