+++ crates/iceberg/src/expr/visitors/mod.rs
@@ -1,1 +1,3 @@
-mod bound_predicate_evaluator;
+pub(crate) mod bound_predicate_evaluator;
+pub(crate) mod inclusive_projection;
+pub(crate) mod manifest_evaluator;
