+++ crates/iceberg/src/scan.rs
@@ -18,13 +18,10 @@
 //! Table scan api.
 
 use crate::arrow::ArrowReaderBuilder;
-use crate::expr::BoundPredicate::AlwaysTrue;
-use crate::expr::{Bind, BoundPredicate, Predicate, PredicateOperator};
+use crate::expr::visitors::manifest_evaluator::ManifestEvaluatorFactory;
+use crate::expr::{Bind, Predicate};
 use crate::io::FileIO;
-use crate::spec::{
-    DataContentType, FieldSummary, ManifestEntryRef, ManifestFile, PartitionField,
-    PartitionSpecRef, Schema, SchemaRef, SnapshotRef, TableMetadataRef,
-};
+use crate::spec::{DataContentType, ManifestEntryRef, SchemaRef, SnapshotRef, TableMetadataRef};
 use crate::table::Table;
 use crate::{Error, ErrorKind};
 use arrow_array::RecordBatch;
@@ -32,7 +29,6 @@ use async_stream::try_stream;
 use futures::stream::{iter, BoxStream};
 use futures::StreamExt;
 use std::collections::HashMap;
-use std::sync::Arc;
 
 /// Builder to create table scan.
 pub struct TableScanBuilder<'a> {
@@ -174,8 +170,9 @@ impl TableScan {
     /// Returns a stream of file scan tasks.
 
     pub async fn plan_files(&'static self) -> crate::Result<FileScanTaskStream> {
-        // Cache `PartitionEvaluator`s created as part of this scan
-        let mut manifest_evaluator_cache: HashMap<i32, ManifestEvaluator> = HashMap::new();
+        // Cache `ManifestEvaluatorFactory`s created as part of this scan
+        let mut manifest_evaluator_factory_cache: HashMap<i32, ManifestEvaluatorFactory> =
+            HashMap::new();
 
         let snapshot = self.snapshot.clone();
         let table_metadata = self.table_metadata.clone();
@@ -194,13 +191,13 @@ impl TableScan {
                 // Use one from the cache if there is one. If not, create one, put it in
                 // the cache, and take a reference to it.
                 if let Some(filter) = self.filter.as_ref() {
-                    let manifest_evaluator = manifest_evaluator_cache
+                    let manifest_eval_factory = manifest_evaluator_factory_cache
                             .entry(entry.partition_spec_id())
-                            .or_insert_with_key(|key| self.create_manifest_evaluator(key, filter));
+                            .or_insert_with_key(|key| self.create_manifest_eval_factory(key, filter));
 
 
                     // reject any manifest files whose partition values don't match the filter.
-                    if !manifest_evaluator.eval(entry) {
+                    if !manifest_eval_factory.evaluate(entry)? {
                         continue;
                     }
                 }
@@ -231,14 +228,18 @@ impl TableScan {
         .boxed())
     }
 
-    fn create_manifest_evaluator(&self, id: &i32, filter: &Predicate) -> ManifestEvaluator {
+    fn create_manifest_eval_factory(
+        &self,
+        id: &i32,
+        filter: &Predicate,
+    ) -> ManifestEvaluatorFactory {
         let bound_predicate = filter
             .bind(self.schema.clone(), self.case_sensitive)
             .unwrap();
 
         let partition_spec = self.table_metadata.partition_spec_by_id(*id).unwrap();
 
-        ManifestEvaluator::new(
+        ManifestEvaluatorFactory::new(
             partition_spec.clone(),
             self.schema.clone(),
             bound_predicate,
@@ -275,207 +276,6 @@ pub type ArrowRecordBatchStream = BoxStream<'static, crate::Result<RecordBatch>>
 impl FileScanTask {
     pub fn data_file(&self) -> ManifestEntryRef {
         self.data_file.clone()
-    }
-}
-
-struct ManifestEvaluator {
-    #[allow(dead_code)]
-    partition_schema: SchemaRef,
-    partition_filter: BoundPredicate,
-    #[allow(dead_code)]
-    case_sensitive: bool,
-}
-
-impl ManifestEvaluator {
-    pub(crate) fn new(
-        partition_spec: PartitionSpecRef,
-        table_schema: SchemaRef,
-        partition_filter: BoundPredicate,
-        case_sensitive: bool,
-    ) -> crate::Result<Self> {
-        let partition_type = partition_spec.partition_type(&table_schema)?;
-
-        // this is needed as SchemaBuilder.with_fields expects an iterator over
-        // Arc<NestedField> rather than &Arc<NestedField>
-        let cloned_partition_fields: Vec<_> =
-            partition_type.fields().iter().map(Arc::clone).collect();
-
-        let partition_schema = Schema::builder()
-            .with_fields(cloned_partition_fields)
-            .build()?;
-
-        let partition_schema_ref = Arc::new(partition_schema);
-
-        let inclusive_projection =
-            InclusiveProjection::new(table_schema.clone(), partition_spec.clone());
-        let unbound_partition_filter = inclusive_projection.project(&partition_filter)?;
-
-        let partition_filter =
-            unbound_partition_filter.bind(partition_schema_ref.clone(), case_sensitive)?;
-
-        Ok(Self {
-            partition_schema: partition_schema_ref,
-            partition_filter,
-            case_sensitive,
-        })
-    }
-
-    pub(crate) fn eval(&self, manifest_file: &ManifestFile) -> bool {
-        if manifest_file.partitions.is_empty() {
-            return true;
-        }
-
-        self.visit(&self.partition_filter, &manifest_file.partitions)
-    }
-
-    // see https://github.com/apache/iceberg-python/blob/ea9da8856a686eaeda0d5c2be78d5e3102b67c44/pyiceberg/expressions/visitors.py#L548
-    fn visit(&self, predicate: &BoundPredicate, partitions: &Vec<FieldSummary>) -> bool {
-        match predicate {
-            AlwaysTrue => true,
-            BoundPredicate::AlwaysFalse => false,
-            BoundPredicate::And(expr) => {
-                self.visit(expr.inputs()[0], partitions) && self.visit(expr.inputs()[1], partitions)
-            }
-            BoundPredicate::Or(expr) => {
-                self.visit(expr.inputs()[0], partitions) || self.visit(expr.inputs()[1], partitions)
-            }
-            BoundPredicate::Not(_) => {
-                panic!("NOT predicates should be eliminated before calling this function")
-            }
-            BoundPredicate::Unary(expr) => {
-                let pos = expr.term().accessor().position();
-                let field = &partitions[pos as usize];
-
-                match expr.op() {
-                    PredicateOperator::IsNull => field.contains_null,
-                    PredicateOperator::NotNull => {
-                        todo!()
-                    }
-                    PredicateOperator::IsNan => field.contains_nan.is_some(),
-                    PredicateOperator::NotNan => {
-                        todo!()
-                    }
-                    _ => {
-                        panic!("unexpected op")
-                    }
-                }
-            }
-            BoundPredicate::Binary(expr) => {
-                let pos = expr.term().accessor().position();
-                let _field = &partitions[pos as usize];
-
-                match expr.op() {
-                    PredicateOperator::LessThan => {
-                        todo!()
-                    }
-                    PredicateOperator::LessThanOrEq => {
-                        todo!()
-                    }
-                    PredicateOperator::GreaterThan => {
-                        todo!()
-                    }
-                    PredicateOperator::GreaterThanOrEq => {
-                        todo!()
-                    }
-                    PredicateOperator::Eq => {
-                        todo!()
-                    }
-                    PredicateOperator::NotEq => {
-                        todo!()
-                    }
-                    PredicateOperator::StartsWith => {
-                        todo!()
-                    }
-                    PredicateOperator::NotStartsWith => {
-                        todo!()
-                    }
-                    _ => {
-                        panic!("unexpected op")
-                    }
-                }
-            }
-            BoundPredicate::Set(expr) => {
-                let pos = expr.term().accessor().position();
-                let _field = &partitions[pos as usize];
-
-                match expr.op() {
-                    PredicateOperator::In => {
-                        todo!()
-                    }
-                    PredicateOperator::NotIn => true,
-                    _ => {
-                        panic!("unexpected op")
-                    }
-                }
-            }
-        }
-    }
-}
-
-struct InclusiveProjection {
-    #[allow(dead_code)]
-    table_schema: SchemaRef,
-    partition_spec: PartitionSpecRef,
-}
-
-impl InclusiveProjection {
-    pub(crate) fn new(table_schema: SchemaRef, partition_spec: PartitionSpecRef) -> Self {
-        Self {
-            table_schema,
-            partition_spec,
-        }
-    }
-
-    pub(crate) fn project(&self, predicate: &BoundPredicate) -> crate::Result<Predicate> {
-        self.visit(predicate)
-    }
-
-    fn visit(&self, bound_predicate: &BoundPredicate) -> crate::Result<Predicate> {
-        Ok(match bound_predicate {
-            BoundPredicate::AlwaysTrue => Predicate::AlwaysTrue,
-            BoundPredicate::AlwaysFalse => Predicate::AlwaysFalse,
-            BoundPredicate::And(expr) => {
-                let [left_pred, right_pred] = expr.inputs();
-                self.visit(left_pred)?.and(self.visit(right_pred)?)
-            }
-            BoundPredicate::Or(expr) => {
-                let [left_pred, right_pred] = expr.inputs();
-                self.visit(left_pred)?.or(self.visit(right_pred)?)
-            }
-            BoundPredicate::Not(_) => {
-                panic!("should not get here as NOT-rewriting should have removed NOT nodes")
-            }
-            bp => self.visit_bound_predicate(bp)?,
-        })
-    }
-
-    fn visit_bound_predicate(&self, predicate: &BoundPredicate) -> crate::Result<Predicate> {
-        let field_id = match predicate {
-            BoundPredicate::Unary(expr) => expr.field_id(),
-            BoundPredicate::Binary(expr) => expr.field_id(),
-            BoundPredicate::Set(expr) => expr.field_id(),
-            _ => {
-                panic!("Should not get here as these branches handled in self.visit")
-            }
-        };
-
-        // TODO: cache this?
-        let mut parts: Vec<&PartitionField> = vec![];
-        for partition_spec_field in &self.partition_spec.fields {
-            if partition_spec_field.source_id == field_id {
-                parts.push(partition_spec_field)
-            }
-        }
-
-        parts.iter().fold(Ok(Predicate::AlwaysTrue), |res, &part| {
-            Ok(
-                if let Some(pred_for_part) = part.transform.project(&part.name, predicate)? {
-                    res?.and(pred_for_part)
-                } else {
-                    res?
-                },
-            )
-        })
     }
 }
